간단 댓글 구현

+++ src/boards/boards.service.ts
@@ -12,8 +12,8 @@ export class BoardsService {
   constructor(private readonly boardRepository: BoardsRepository) {}
 
   async findAll(): Promise<Board[]> {
-    const status = BoardStatus.PUBLIC
-    return await this.boardRepository.find({where : {status}});
+    const status = BoardStatus.PUBLIC;
+    return await this.boardRepository.find({ where: { status } });
   }
   async findById(id: number): Promise<Board> {
     const found = await this.boardRepository.findOne({ where: { id } });
@@ -42,26 +42,17 @@ export class BoardsService {
     return await this.boardRepository.createBoard(createBaordDto, user);
   }
   async update(id: number, updateBoardDto: UpdateBoardDto, user: User): Promise<boolean> {
-    const result = await this.boardRepository.createQueryBuilder('board')
-      .update()
-      .set(updateBoardDto)
-      .where("id=:id AND userId = :userId", { id, userId: user.id })
-      .execute();
+    const result = await this.boardRepository.createQueryBuilder("board").update().set(updateBoardDto).where("id=:id AND userId = :userId", { id, userId: user.id }).execute();
     return result.affected > 0;
   }
-  async updateStatus(id: number, status: BoardStatus, user:User): Promise<Board> {
-    const board = await this.boardRepository.createQueryBuilder('board')
-      .where('id = :id AND userId = :userId', { id, userId: user.id })
-      .getOne();
+  async updateStatus(id: number, status: BoardStatus, user: User): Promise<Board> {
+    const board = await this.boardRepository.createQueryBuilder("board").where("id = :id AND userId = :userId", { id, userId: user.id }).getOne();
     board.status = status;
     const result = await this.boardRepository.save(board);
     return result;
   }
   async delete(id: number, user: User): Promise<DeleteResult> {
-    const result = await this.boardRepository.createQueryBuilder('board')
-    .delete()
-    .where("id=:id AND userId = :userId", {id, userId : user.id})
-    .execute();
+    const result = await this.boardRepository.createQueryBuilder("board").delete().where("id=:id AND userId = :userId", { id, userId: user.id }).execute();
 
     if (result.affected === 0) throw new NotFoundException(`Can't find ${id}`);
